Appearence and Reload Cell
- Appearence Model created
- Resize opened cell method created

+++ KMAccordionTableViewController/Classes/Model/KMSection.h
@@ -17,6 +17,6 @@
 @property UIView *overHeaderView;
 @property KMSectionHeaderView *headerView;
 @property(nonatomic, copy) NSString *title;
-@property int sectionIndex;
+@property NSInteger sectionIndex;
 
 @end
